+++ pkgs/ffigen/lib/src/code_generator/func.dart
@@ -111,7 +111,6 @@ class Func extends LookUpBinding {
         functionType.getFfiDartType(w, writeArgumentNames: false);
     final needsWrapper = !functionType.sameDartAndFfiDartType && !isInternal;
 
-    final isLeafString = isLeaf ? 'isLeaf:true' : '';
     final funcVarName = w.wrapperLevelUniqueNamer.makeUnique('_$name');
     final ffiReturnType = functionType.returnType.getFfiDartType(w);
     final ffiArgDeclString = functionType.dartTypeParameters
@@ -146,12 +145,15 @@ class Func extends LookUpBinding {
     }
 
     if (ffiNativeConfig.enabled) {
-      final assetString = ffiNativeConfig.assetId != null
-          ? ", assetId: '${ffiNativeConfig.assetId}'"
-          : '';
       final nativeFuncName = needsWrapper ? funcVarName : enclosingFuncName;
       s.write('''
-@${w.ffiLibraryPrefix}.Native<$cType>(symbol: '$originalName'$assetString$isLeafString)
+${makeNativeAnnotation(
+        w,
+        nativeType: cType,
+        dartName: nativeFuncName,
+        nativeSymbolName: originalName,
+        isLeaf: isLeaf,
+      )}
 external $ffiReturnType $nativeFuncName($ffiArgDeclString);
 
 ''');
@@ -164,8 +166,18 @@ $dartReturnType $enclosingFuncName($libArg$dartArgDeclString) => $funcImplCall;
 
 ''');
       }
+
+      if (exposeSymbolAddress) {
+        // Add to SymbolAddress in writer.
+        w.symbolAddressWriter.addNativeSymbol(
+          type:
+              '${w.ffiLibraryPrefix}.Pointer<${w.ffiLibraryPrefix}.NativeFunction<$cType>>',
+          name: name,
+        );
+      }
     } else {
       funcPointerName = w.wrapperLevelUniqueNamer.makeUnique('_${name}Ptr');
+      final isLeafString = isLeaf ? 'isLeaf:true' : '';
 
       // Write enclosing function.
       s.write('''