+++ pkgs/ffigen/lib/src/code_generator/utils.dart
@@ -1,4 +1,11 @@
+// Copyright (c) 2023, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 import 'dart_keywords.dart';
+import 'pointer.dart';
+import 'type.dart';
+import 'writer.dart';
 
 class UniqueNamer {
   final Set<String> _usedUpNames;
@@ -74,4 +81,34 @@ String makeDoc(String text) {
   s.write('\n');
 
   return s.toString();
+}
+
+String makeNativeAnnotation(
+  Writer w, {
+  required String? nativeType,
+  required String dartName,
+  required String nativeSymbolName,
+  bool isLeaf = false,
+}) {
+  final args = <(String, String)>[];
+  if (dartName != nativeSymbolName) {
+    args.add(('symbol', '"$nativeSymbolName"'));
+  }
+  if (isLeaf) {
+    args.add(('isLeaf', 'true'));
+  }
+
+  final combinedArgs = args.map((e) => '${e.$1}: ${e.$2}').join(', ');
+  return '@${w.ffiLibraryPrefix}.Native<$nativeType>($combinedArgs)';
+}
+
+String makeArrayAnnotation(Writer w, ConstantArray arrayType) {
+  final dimensions = <int>[];
+  Type type = arrayType;
+  while (type is ConstantArray) {
+    dimensions.add(type.length);
+    type = type.child;
+  }
+
+  return '@${w.ffiLibraryPrefix}.Array.multi([${dimensions.join(', ')}])';
 }