+++ pkgs/native_assets_cli/lib/src/model/asset.dart
@@ -5,19 +5,17 @@
 import 'package:collection/collection.dart';
 import 'package:yaml/yaml.dart';
 
-import '../utils/uri.dart';
+import '../api/asset.dart' as api;
 import '../utils/yaml.dart';
 import 'build_type.dart';
 import 'link_mode.dart';
 import 'target.dart';
 
-abstract class AssetPath {
+abstract class AssetPath implements api.AssetPath {
   factory AssetPath(String pathType, Uri? uri) {
     switch (pathType) {
       case AssetAbsolutePath._pathTypeValue:
         return AssetAbsolutePath(uri!);
-      case AssetRelativePath._pathTypeValue:
-        return AssetRelativePath(uri!);
       case AssetSystemPath._pathTypeValue:
         return AssetSystemPath(uri!);
       case AssetInExecutable._pathTypeValue:
@@ -36,16 +34,14 @@ abstract class AssetPath {
   }
 
   Map<String, Object> toYaml();
-  List<String> toDartConst();
 
   static const _pathTypeKey = 'path_type';
   static const _uriKey = 'uri';
-
-  Future<bool> exists();
 }
 
 /// Asset at absolute path [uri].
-class AssetAbsolutePath implements AssetPath {
+class AssetAbsolutePath implements AssetPath, api.AssetAbsolutePath {
+  @override
   final Uri uri;
 
   AssetAbsolutePath(this.uri);
@@ -59,9 +55,6 @@ class AssetAbsolutePath implements AssetPath {
       };
 
   @override
-  List<String> toDartConst() => [_pathTypeValue, uri.toFilePath()];
-
-  @override
   int get hashCode => Object.hash(uri, 133711);
 
   @override
@@ -71,50 +64,13 @@ class AssetAbsolutePath implements AssetPath {
     }
     return uri == other.uri;
   }
-
-  @override
-  Future<bool> exists() => uri.fileSystemEntity.exists();
-}
-
-/// Asset is avaliable on a relative path.
-///
-/// If [LinkMode] of an [Asset] is [LinkMode.dynamic],
-/// `Platform.script.resolve(uri)` will be used to load the asset at runtime.
-class AssetRelativePath implements AssetPath {
-  final Uri uri;
-
-  AssetRelativePath(this.uri);
-
-  static const _pathTypeValue = 'relative';
-
-  @override
-  Map<String, Object> toYaml() => {
-        AssetPath._pathTypeKey: _pathTypeValue,
-        AssetPath._uriKey: uri.toFilePath(),
-      };
-
-  @override
-  List<String> toDartConst() => [_pathTypeValue, uri.toFilePath()];
-
-  @override
-  int get hashCode => Object.hash(uri, 133717);
-
-  @override
-  bool operator ==(Object other) {
-    if (other is! AssetRelativePath) {
-      return false;
-    }
-    return uri == other.uri;
-  }
-
-  @override
-  Future<bool> exists() => uri.fileSystemEntity.exists();
 }
 
 /// Asset is avaliable on the system `PATH`.
 ///
 /// [uri] only contains a file name.
-class AssetSystemPath implements AssetPath {
+class AssetSystemPath implements AssetPath, api.AssetSystemPath {
+  @override
   final Uri uri;
 
   AssetSystemPath(this.uri);
@@ -128,9 +84,6 @@ class AssetSystemPath implements AssetPath {
       };
 
   @override
-  List<String> toDartConst() => [_pathTypeValue, uri.toFilePath()];
-
-  @override
   int get hashCode => Object.hash(uri, 133723);
 
   @override
@@ -140,14 +93,11 @@ class AssetSystemPath implements AssetPath {
     }
     return uri == other.uri;
   }
-
-  @override
-  Future<bool> exists() => Future.value(true);
 }
 
 /// Asset is loaded in the process and symbols are available through
 /// `DynamicLibrary.process()`.
-class AssetInProcess implements AssetPath {
+class AssetInProcess implements AssetPath, api.AssetInProcess {
   AssetInProcess._();
 
   static final AssetInProcess _singleton = AssetInProcess._();
@@ -160,17 +110,11 @@ class AssetInProcess implements AssetPath {
   Map<String, Object> toYaml() => {
         AssetPath._pathTypeKey: _pathTypeValue,
       };
-
-  @override
-  List<String> toDartConst() => [_pathTypeValue];
-
-  @override
-  Future<bool> exists() => Future.value(true);
 }
 
 /// Asset is embedded in executable and symbols are available through
 /// `DynamicLibrary.executable()`.
-class AssetInExecutable implements AssetPath {
+class AssetInExecutable implements AssetPath, api.AssetInExecutable {
   AssetInExecutable._();
 
   static final AssetInExecutable _singleton = AssetInExecutable._();
@@ -183,18 +127,16 @@ class AssetInExecutable implements AssetPath {
   Map<String, Object> toYaml() => {
         AssetPath._pathTypeKey: _pathTypeValue,
       };
-
-  @override
-  List<String> toDartConst() => [_pathTypeValue];
-
-  @override
-  Future<bool> exists() => Future.value(true);
 }
 
-class Asset {
+class Asset implements api.Asset {
+  @override
   final LinkMode linkMode;
+  @override
   final String id;
+  @override
   final Target target;
+  @override
   final AssetPath path;
 
   /// The step at which the asset should be written to file.
@@ -277,19 +219,13 @@ class Asset {
         _step: step.name,
       };
 
-  Map<String, List<String>> toDartConst() => {
-        id: path.toDartConst(),
-      };
-
-  String toYamlString() => yamlEncode(toYaml());
-
   static const _idKey = 'id';
   static const _linkModeKey = 'link_mode';
   static const _pathKey = 'path';
   static const _targetKey = 'target';
   static const _step = 'step';
 
-  Future<bool> exists() => path.exists();
+  // Future<bool> exists() => path.exists();
 
   @override
   String toString() => 'Asset(${toYaml()})';
@@ -299,44 +235,4 @@ extension AssetIterable on Iterable<Asset> {
   List<Object> toYaml() => [for (final item in this) item.toYaml()];
 
   String toYamlString() => yamlEncode(toYaml());
-
-  Iterable<Asset> whereLinkMode(LinkMode linkMode) =>
-      where((e) => e.linkMode == linkMode);
-
-  Map<Target, List<Asset>> get assetsPerTarget {
-    final result = <Target, List<Asset>>{};
-    for (final asset in this) {
-      final assets = result[asset.target] ?? [];
-      assets.add(asset);
-      result[asset.target] = assets;
-    }
-    return result;
-  }
-
-  Map<String, Map<String, List<String>>> toDartConst() => {
-        for (final entry in assetsPerTarget.entries)
-          entry.key.toString():
-              _combineMaps(entry.value.map((e) => e.toDartConst()).toList())
-      };
-
-  Map<Object, Object> toNativeAssetsFileEncoding() => {
-        'format-version': [1, 0, 0],
-        'native-assets': toYaml(),
-      };
-
-  String toNativeAssetsFile() => yamlEncode(toNativeAssetsFileEncoding());
-
-  Future<bool> allExist() async {
-    final allResults = await Future.wait(map((e) => e.exists()));
-    final missing = allResults.contains(false);
-    return !missing;
-  }
-}
-
-Map<X, Y> _combineMaps<X, Y>(Iterable<Map<X, Y>> maps) {
-  final result = <X, Y>{};
-  for (final map in maps) {
-    result.addAll(map);
-  }
-  return result;
 }