+++ pkgs/native_assets_cli/test/helpers.dart
@@ -4,7 +4,8 @@
 
 import 'dart:io';
 
-import 'package:native_assets_cli/src/model/build_config.dart';
+import 'package:native_assets_cli/native_assets_cli.dart';
+import 'package:native_assets_cli/native_assets_cli_internal.dart' as internal;
 
 const keepTempKey = 'KEEP_TEMPORARY_DIRECTORIES';
 
@@ -74,32 +75,67 @@ extension on Uri {
 String unparseKey(String key) => key.replaceAll('.', '__').toUpperCase();
 
 /// Archiver provided by the environment.
+///
+/// Provided on Dart CI.
 final Uri? ar = Platform
-    .environment[unparseKey(CCompilerConfig.arConfigKeyFull)]
+    .environment[unparseKey(internal.CCompilerConfig.arConfigKeyFull)]
     ?.asFileUri();
 
 /// Compiler provided by the environment.
+///
+/// Provided on Dart CI.
 final Uri? cc = Platform
-    .environment[unparseKey(CCompilerConfig.ccConfigKeyFull)]
+    .environment[unparseKey(internal.CCompilerConfig.ccConfigKeyFull)]
     ?.asFileUri();
 
 /// Linker provided by the environment.
+///
+/// Provided on Dart CI.
 final Uri? ld = Platform
-    .environment[unparseKey(CCompilerConfig.ldConfigKeyFull)]
+    .environment[unparseKey(internal.CCompilerConfig.ldConfigKeyFull)]
     ?.asFileUri();
 
 /// Path to script that sets environment variables for [cc], [ld], and [ar].
 ///
-/// Provided by environment.
+/// Provided on Dart CI.
 final Uri? envScript = Platform
-    .environment[unparseKey(CCompilerConfig.envScriptConfigKeyFull)]
+    .environment[unparseKey(internal.CCompilerConfig.envScriptConfigKeyFull)]
     ?.asFileUri();
 
 /// Arguments for [envScript] provided by environment.
-final List<String>? envScriptArgs = Platform
-    .environment[unparseKey(CCompilerConfig.envScriptArgsConfigKeyFull)]
+///
+/// Provided on Dart CI.
+final List<String>? envScriptArgs = Platform.environment[
+        unparseKey(internal.CCompilerConfig.envScriptArgsConfigKeyFull)]
     ?.split(' ');
 
 extension on String {
   Uri asFileUri() => Uri.file(this);
+}
+
+extension AssetIterable on Iterable<Asset> {
+  Future<bool> allExist() async {
+    final allResults = await Future.wait(map((e) => e.exists()));
+    final missing = allResults.contains(false);
+    return !missing;
+  }
+}
+
+extension on Asset {
+  Future<bool> exists() async {
+    final path_ = path;
+    return switch (path_) {
+      AssetAbsolutePath _ => await path_.uri.fileSystemEntity.exists(),
+      _ => true,
+    };
+  }
+}
+
+extension UriExtension on Uri {
+  FileSystemEntity get fileSystemEntity {
+    if (path.endsWith(Platform.pathSeparator) || path.endsWith('/')) {
+      return Directory.fromUri(this);
+    }
+    return File.fromUri(this);
+  }
 }