+++ pkgs/native_assets_cli/test/model/build_output_test.dart
@@ -4,7 +4,7 @@
 
 import 'dart:io';
 
-import 'package:native_assets_cli/native_assets_cli.dart';
+import 'package:native_assets_cli/native_assets_cli_internal.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -29,7 +29,7 @@ void main() {
       ),
       Asset(
         id: 'foo2',
-        path: AssetRelativePath(Uri(path: 'path/to/libfoo2.so')),
+        path: AssetSystemPath(Uri(path: 'path/to/libfoo2.so')),
         target: Target.androidX64,
         linkMode: LinkMode.dynamic,
       ),
@@ -54,7 +54,7 @@ assets:
   - id: foo2
     link_mode: dynamic
     path:
-      path_type: relative
+      path_type: system
       uri: path/to/libfoo2.so
     target: android_x64
     type: code