+++ pkgs/native_toolchain_c/lib/src/cbuilder/run_cbuilder.dart
@@ -232,7 +232,8 @@ class RunCBuilder {
   Future<void> runCl({required ToolInstance compiler}) async {
     final vcvars = (await _resolver.toolchainEnvironmentScript(compiler))!;
     final vcvarsArgs = _resolver.toolchainEnvironmentScriptArguments();
-    final environment = await envFromBat(vcvars, arguments: vcvarsArgs ?? []);
+    final environment =
+        await environmentFromBatchFile(vcvars, arguments: vcvarsArgs ?? []);
 
     final isStaticLib = staticLibrary != null;
     Uri? archiver_;