+++ pkgs/native_toolchain_c/CHANGELOG.md
@@ -1,5 +1,18 @@
-## 0.3.3-wip
+## 0.3.5-wip
 
+- Bump `package:native_assets_cli` to 0.4.1.
+
+## 0.3.4+1
+
+- Stop depending on private `package:native_assets_cli` `CCompilerConfig` fields.
+
+## 0.3.4
+
+- Bump `package:native_assets_cli` to 0.4.0.
+
+## 0.3.3
+
+- Export `environmentFromBatchFile`.
 - Bump `package:native_assets_cli` to 0.3.2.
 
 ## 0.3.2